move adapter pattern, from delivery to example
